OCR/requirements.txt.txt
streamlit==1.29.0
pillow==10.0.0
numpy==1.26.5
easyocr==1.7.0
pandas==2.1.1
opencv-python-headless==4.8.1.1
python-dateutil==2.8.2
